+++ aws/logs_monitoring/lambda_function.py
@@ -28,6 +28,32 @@ from enhanced_lambda_metrics import (
     get_enriched_lambda_log_tags,
     parse_and_submit_enhanced_metrics,
 )
+from settings import (
+    DD_API_KEY,
+    DD_FORWARD_LOG,
+    DD_USE_TCP,
+    DD_USE_COMPRESSION,
+    DD_COMPRESSION_LEVEL,
+    DD_NO_SSL,
+    DD_SKIP_SSL_VALIDATION,
+    DD_SITE,
+    DD_TAGS,
+    DD_API_URL,
+    DD_TRACE_INTAKE_URL,
+    DD_URL,
+    DD_PORT,
+    SCRUBBING_RULE_CONFIGS,
+    INCLUDE_AT_MATCH,
+    EXCLUDE_AT_MATCH,
+    DD_MULTILINE_LOG_REGEX_PATTERN,
+    multiline_regex,
+    multiline_regex_start_pattern,
+    DD_SOURCE,
+    DD_CUSTOM_TAGS,
+    DD_SERVICE,
+    DD_HOST,
+    DD_FORWARDER_VERSION,
+)
 
 
 log = logging.getLogger()
@@ -35,6 +61,33 @@ log.setLevel(logging.getLevelName(os.environ.get("DD_LOG_LEVEL", "INFO").upper()
 
 DD_FORWARD_TRACES = True
 DD_FORWARD_METRIC = True
+
+log.debug(
+    DD_API_KEY,
+    DD_FORWARD_LOG,
+    DD_USE_TCP,
+    DD_USE_COMPRESSION,
+    DD_COMPRESSION_LEVEL,
+    DD_NO_SSL,
+    DD_SKIP_SSL_VALIDATION,
+    DD_SITE,
+    DD_TAGS,
+    DD_API_URL,
+    DD_TRACE_INTAKE_URL,
+    DD_URL,
+    DD_PORT,
+    SCRUBBING_RULE_CONFIGS,
+    INCLUDE_AT_MATCH,
+    EXCLUDE_AT_MATCH,
+    DD_MULTILINE_LOG_REGEX_PATTERN,
+    multiline_regex,
+    multiline_regex_start_pattern,
+    DD_SOURCE,
+    DD_CUSTOM_TAGS,
+    DD_SERVICE,
+    DD_HOST,
+    DD_FORWARDER_VERSION,
+)
 
 try:
     import requests
@@ -48,162 +101,29 @@ except ImportError:
     # of requests is removed in botocore 1.13.x.
     from botocore.vendored import requests
 
-
-def get_env_var(envvar, default, boolean=False):
-    """
-        Return the value of the given environment variable with debug logging.
-        When boolean=True, parse the value as a boolean case-insensitively.
-    """
-    value = os.getenv(envvar, default=default)
-    if boolean:
-        value = value.lower() == "true"
-    log.debug(f"{envvar}: {value}")
-    return value
-
-
-#####################################
-############# PARAMETERS ############
-#####################################
-
-## @param DD_API_KEY - String - conditional - default: none
-## The Datadog API key associated with your Datadog Account
-## It can be found here:
-##
-##   * Datadog US Site: https://app.datadoghq.com/account/settings#api
-##   * Datadog EU Site: https://app.datadoghq.eu/account/settings#api
-##
-## Must be set if one of the following is not set: DD_API_KEY_SECRET_ARN, DD_API_KEY_SSM_NAME, DD_KMS_API_KEY
-#
-DD_API_KEY = "<YOUR_DATADOG_API_KEY>"
-
-## @param DD_API_KEY_SECRET_ARN - String - optional - default: none
-## ARN of Datadog API key stored in AWS Secrets Manager
-##
-## Supercedes: DD_API_KEY_SSM_NAME, DD_KMS_API_KEY, DD_API_KEY
-
-## @param DD_API_KEY_SSM_NAME - String - optional - default: none
-## Name of parameter containing Datadog API key in AWS SSM Parameter Store
-##
-## Supercedes: DD_KMS_API_KEY, DD_API_KEY
-
-## @param DD_KMS_API_KEY - String - optional - default: none
-## AWS KMS encrypted Datadog API key
-##
-## Supercedes: DD_API_KEY
-
-## @param DD_FORWARD_LOG - boolean - optional - default: true
-## Set this variable to `False` to disable log forwarding.
-## E.g., when you only want to forward metrics from logs.
-#
-DD_FORWARD_LOG = get_env_var("DD_FORWARD_LOG", "true", boolean=True)
-
-## @param DD_USE_TCP - boolean - optional -default: false
-## Change this value to `true` to send your logs and metrics using the TCP network client
-## By default, it uses the HTTP client.
-#
-DD_USE_TCP = get_env_var("DD_USE_TCP", "false", boolean=True)
-
-## @param DD_USE_COMPRESSION - boolean - optional -default: true
-## Only valid when sending logs over HTTP
-## Change this value to `false` to send your logs without any compression applied
-## By default, compression is enabled.
-#
-DD_USE_COMPRESSION = get_env_var("DD_USE_COMPRESSION", "true", boolean=True)
-
-## @param DD_USE_COMPRESSION - integer - optional -default: 6
-## Change this value to set the compression level.
-## Values range from 0 (no compression) to 9 (best compression).
-## By default, compression is set to level 6.
-#
-DD_COMPRESSION_LEVEL = int(os.getenv("DD_COMPRESSION_LEVEL", 6))
-
-## @param DD_USE_SSL - boolean - optional -default: false
-## Change this value to `true` to disable SSL
-## Useful when you are forwarding your logs to a proxy.
-#
-DD_NO_SSL = get_env_var("DD_NO_SSL", "false", boolean=True)
-
-## @param DD_SKIP_SSL_VALIDATION - boolean - optional -default: false
-## Disable SSL certificate validation when forwarding logs via HTTP.
-#
-DD_SKIP_SSL_VALIDATION = get_env_var("DD_SKIP_SSL_VALIDATION", "false", boolean=True)
-
-## @param DD_SITE - String - optional -default: datadoghq.com
-## Define the Datadog Site to send your logs and metrics to.
-## Set it to `datadoghq.eu` to send your logs and metrics to Datadog EU site.
-#
-DD_SITE = get_env_var("DD_SITE", default="datadoghq.com")
-
-## @param DD_TAGS - list of comma separated strings - optional -default: none
-## Pass custom tags as environment variable or through this variable.
-## Ensure your tags are a comma separated list of strings with no trailing comma in the envvar!
-#
-DD_TAGS = get_env_var("DD_TAGS", "")
-
-## @param DD_API_URL - Url to use for  validating the the api key. Used for validating api key.
-DD_API_URL = get_env_var("DD_API_URL", default="https://api.{}".format(DD_SITE))
-log.debug(f"DD_API_URL: {DD_API_URL}")
-
-## @param DD_TRACE_INTAKE_URL - Url to use for  validating the the api key. Used for validating api key.
-DD_TRACE_INTAKE_URL = get_env_var(
-    "DD_TRACE_INTAKE_URL", default="https://trace.agent.{}".format(DD_SITE)
+# DD_API_KEY must be set
+if DD_API_KEY == "<YOUR_DATADOG_API_KEY>" or DD_API_KEY == "":
+    raise Exception("Missing Datadog API key")
+# Check if the API key is the correct number of characters
+if len(DD_API_KEY) != 32:
+    raise Exception(
+        "The API key is not the expected length. "
+        "Please confirm that your API key is correct"
+    )
+# Validate the API key
+validation_res = requests.get(
+    "{}/api/v1/validate?api_key={}".format(DD_API_URL, DD_API_KEY)
 )
+if not validation_res.ok:
+    raise Exception("The API key is not valid.")
 
-if DD_USE_TCP:
-    DD_URL = get_env_var("DD_URL", default="lambda-intake.logs." + DD_SITE)
-    try:
-        if "DD_SITE" in os.environ and DD_SITE == "datadoghq.eu":
-            DD_PORT = int(get_env_var("DD_PORT", default="443"))
-        else:
-            DD_PORT = int(get_env_var("DD_PORT", default="10516"))
-    except Exception:
-        DD_PORT = 10516
-else:
-    DD_URL = get_env_var("DD_URL", default="lambda-http-intake.logs." + DD_SITE)
-    DD_PORT = int(get_env_var("DD_PORT", default="443"))
+# Force the layer to use the exact same API key and host as the forwarder
+api._api_key = DD_API_KEY
+api._api_host = DD_API_URL
 
-## @param DD_USE_PRIVATE_LINK - whether to forward logs via private link
-## Overrides incompatible settings
-#
-DD_USE_PRIVATE_LINK = get_env_var("DD_USE_PRIVATE_LINK", "false", boolean=True)
-if DD_USE_PRIVATE_LINK:
-    log.debug("Private link enabled, overriding configuration settings")
-    # TCP isn't supported when private link is enabled
-    DD_USE_TCP = False
-    DD_NO_SSL = False
-    DD_PORT = 443
-    # Traces aren't supported via private link yet
-    DD_FORWARD_TRACES = False
-    # Override urls to use the private link url
-    DD_URL = "api-pvtlink.logs.datadoghq.com"
-    DD_API_URL = "https://pvtlink.api.datadoghq.com"
-
-
-class ScrubbingRuleConfig(object):
-    def __init__(self, name, pattern, placeholder):
-        self.name = name
-        self.pattern = pattern
-        self.placeholder = placeholder
-
-
-# Scrubbing sensitive data
-# Option to redact all pattern that looks like an ip address / email address / custom pattern
-SCRUBBING_RULE_CONFIGS = [
-    ScrubbingRuleConfig(
-        "REDACT_IP", "\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", "xxx.xxx.xxx.xxx"
-    ),
-    ScrubbingRuleConfig(
-        "REDACT_EMAIL",
-        "[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+",
-        "xxxxx@xxxxx.com",
-    ),
-    ScrubbingRuleConfig(
-        "DD_SCRUBBING_RULE",
-        get_env_var("DD_SCRUBBING_RULE", default=None),
-        get_env_var("DD_SCRUBBING_RULE_REPLACEMENT", default="xxxxx"),
-    ),
-]
-
+trace_connection = None
+if DD_FORWARD_TRACES:
+    trace_connection = TraceConnection(DD_TRACE_INTAKE_URL, DD_API_KEY)
 
 # Use for include, exclude, and scrubbing rules
 def compileRegex(rule, pattern):
@@ -223,88 +143,11 @@ def compileRegex(rule, pattern):
             )
 
 
-# Filtering logs
-# Option to include or exclude logs based on a pattern match
-INCLUDE_AT_MATCH = get_env_var("INCLUDE_AT_MATCH", default=None)
 include_regex = compileRegex("INCLUDE_AT_MATCH", INCLUDE_AT_MATCH)
 
-EXCLUDE_AT_MATCH = get_env_var("EXCLUDE_AT_MATCH", default=None)
 exclude_regex = compileRegex("EXCLUDE_AT_MATCH", EXCLUDE_AT_MATCH)
 
-if "DD_API_KEY_SECRET_ARN" in os.environ:
-    SECRET_ARN = os.environ["DD_API_KEY_SECRET_ARN"]
-    DD_API_KEY = boto3.client("secretsmanager").get_secret_value(SecretId=SECRET_ARN)[
-        "SecretString"
-    ]
-elif "DD_API_KEY_SSM_NAME" in os.environ:
-    SECRET_NAME = os.environ["DD_API_KEY_SSM_NAME"]
-    DD_API_KEY = boto3.client("ssm").get_parameter(
-        Name=SECRET_NAME, WithDecryption=True
-    )["Parameter"]["Value"]
-elif "DD_KMS_API_KEY" in os.environ:
-    ENCRYPTED = os.environ["DD_KMS_API_KEY"]
-    DD_API_KEY = boto3.client("kms").decrypt(
-        CiphertextBlob=base64.b64decode(ENCRYPTED)
-    )["Plaintext"]
-    if type(DD_API_KEY) is bytes:
-        DD_API_KEY = DD_API_KEY.decode("utf-8")
-elif "DD_API_KEY" in os.environ:
-    DD_API_KEY = os.environ["DD_API_KEY"]
-
-# Strip any trailing and leading whitespace from the API key
-DD_API_KEY = DD_API_KEY.strip()
-os.environ["DD_API_KEY"] = DD_API_KEY
-
-# Force the layer to use the exact same API key and host as the forwarder
-api._api_key = DD_API_KEY
-api._api_host = DD_API_URL
-
-# DD_API_KEY must be set
-if DD_API_KEY == "<YOUR_DATADOG_API_KEY>" or DD_API_KEY == "":
-    raise Exception("Missing Datadog API key")
-# Check if the API key is the correct number of characters
-if len(DD_API_KEY) != 32:
-    raise Exception(
-        "The API key is not the expected length. "
-        "Please confirm that your API key is correct"
-    )
-# Validate the API key
-validation_res = requests.get(
-    "{}/api/v1/validate?api_key={}".format(DD_API_URL, DD_API_KEY)
-)
-if not validation_res.ok:
-    raise Exception("The API key is not valid.")
-
-trace_connection = None
-if DD_FORWARD_TRACES:
-    trace_connection = TraceConnection(DD_TRACE_INTAKE_URL, DD_API_KEY)
-
-# DD_MULTILINE_LOG_REGEX_PATTERN: Multiline Log Regular Expression Pattern
-DD_MULTILINE_LOG_REGEX_PATTERN = get_env_var(
-    "DD_MULTILINE_LOG_REGEX_PATTERN", default=None
-)
-if DD_MULTILINE_LOG_REGEX_PATTERN:
-    try:
-        multiline_regex = re.compile(
-            "[\n\r\f]+(?={})".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-        )
-    except Exception:
-        raise Exception(
-            "could not compile multiline regex with pattern: {}".format(
-                DD_MULTILINE_LOG_REGEX_PATTERN
-            )
-        )
-    multiline_regex_start_pattern = re.compile(
-        "^{}".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-    )
-
 rds_regex = re.compile("/aws/rds/(instance|cluster)/(?P<host>[^/]+)/(?P<name>[^/]+)")
-
-DD_SOURCE = "ddsource"
-DD_CUSTOM_TAGS = "ddtags"
-DD_SERVICE = "service"
-DD_HOST = "host"
-DD_FORWARDER_VERSION = "3.15.0"
 
 
 class RetriableException(Exception):
